Separate what LEMMA ships with from what a workspace creates
`runtime_profile_service.py` was 952 lines covering three unrelated things: the
built-in profiles read out of the deployment's environment, the creation of new
ones against a provider that has to be asked first, and the storage and
resolution of both.

* `runtime_system_profiles` -- the profiles nobody created. `system:lemma`
  exists because the deployment has credentials in its environment, which makes
  this the deployment's front door: what these functions read is exactly what a
  new pod can talk to on day one. None of it touches the database.
* `runtime_profile_creation` -- discovery first, then the row. Every creator
  asks the provider before writing, because a profile whose catalog was never
  verified fails later at dispatch with far less context. The round trip runs
  under `connection_released`, since the base URL came from the caller and the
  call is as slow as whatever answers it.
* what is left is storage and resolution, at 506 lines.

`runtime_profile_editor` was reaching three private names across the file
boundary. It still reaches three, but now each one comes from the module that
owns it, which is the difference between a seam and a leak.
